+++ tests/inputs/incorrect-base-class/good_model_form.py
@@ -1,0 +1,5 @@
+from nautobot.apps.forms import NautobotModelForm
+
+
+class MyModelForm(NautobotModelForm):
+    """My model form class."""