+++ tests/inputs/incorrect-base-class/good_primary_model.py
@@ -1,0 +1,5 @@
+from nautobot.apps.models import PrimaryModel
+
+
+class MyModel(PrimaryModel):
+    pass