+++ tests/test_deprecated_status_model.py
@@ -1,15 +1,14 @@
 """Tests for deprecated status model."""
-from pathlib import Path
 
 from pylint.testutils import CheckerTestCase
-from pytest import mark
 
 from pylint_nautobot.deprecated_status_model import NautobotDeprecatedStatusModelChecker
 
 from .utils import assert_error_file
 from .utils import assert_good_file
+from .utils import parametrize_error_files
+from .utils import parametrize_good_files
 
-_INPUTS_PATH = Path(__file__).parent / "inputs/deprecated-status-model/"
 _EXPECTED_ERRORS = {
     "status_model": {
         "msg_id": "nb-status-field-instead-of-status-model",
@@ -25,11 +24,11 @@ class TestDeprecatedStatusModelChecker(CheckerTestCase):
 
     CHECKER_CLASS = NautobotDeprecatedStatusModelChecker
 
-    @mark.parametrize("path", _INPUTS_PATH.glob("error_*.py"))
-    def test_deprecated_status_model(self, path):
-        assert_error_file(self, path, _EXPECTED_ERRORS)
+    @parametrize_error_files(__file__, _EXPECTED_ERRORS)
+    def test_error(self, filename, expected_error):
+        assert_error_file(self, filename, expected_error)
 
     # TBD: Missing test for good_status_model.py
-    @mark.parametrize("path", _INPUTS_PATH.glob("good_*.py"))
-    def test_no_issues(self, path):
-        assert_good_file(self, path)
+    @parametrize_good_files(__file__)
+    def test_good(self, filename):
+        assert_good_file(self, filename)