+++ docs/release_notes.md
@@ -10,6 +10,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-03-05
+
+### Added 
+
+- Added `nb-use-fields-all` rule (#70)
+- Added `nautobot-sub-class-name` rule (#74)
+
+### Changed 
+
+- Improve incorrect base class checker (#72)
+- Update CODEOWNERS (#68)
+
+### Fixed
+
+- Fix RTD docs build (#58)
+- Fix `UIViewSet` sub class name (#76)
+- Fix failing check when class doesNt# have a `Meta` class member (#77)
+
+### Housekeeping
+
+- Allow wider range of `importlib-resources` library (#78)
+
 ## [0.2.1] - 2023-09-01
 
 ### Changed